Выносим всю функциональную часть сортировки из 'Sort' и 'common.js' в 'sorting.js'.
